+++ src/modules/user/controllers/user.controller.ts
@@ -9,7 +9,7 @@ import {
 } from '@nestjs/common';
 import { ParentService } from '../services/parent.service';
 import { Parent } from '../entities/parent.entity';
-import { UpdateUserDto } from '../dto/update-user.dto';
+import { UpdateUserDto } from '../dtos/update-user.dto';
 import { UserType } from '@app-types/module.types';
 
 @Controller('users')
