+++ src/modules/user/dtos/create-user.dto.ts
@@ -1,17 +1,16 @@
 import { UserGender } from '@app-types/module.types';
-import { ApiProperty } from '@nestjs/swagger';
+import { ApiProperty, ApiPropertyOptional } from '@nestjs/swagger';
 import { Type } from 'class-transformer';
 import {
   IsString,
-  IsEmail,
   IsNotEmpty,
-  MinLength,
   IsOptional,
   IsDate,
   IsEnum,
 } from 'class-validator';
+import { ContactInfoDto } from './contact-info.dto';
 
-export class UserBaseDto {
+export class CreateUserDto {
   @IsString()
   @IsNotEmpty()
   @ApiProperty()
@@ -22,20 +21,11 @@ export class UserBaseDto {
   @ApiProperty()
   lastName: string;
 
-  @IsEmail()
   @IsNotEmpty()
   @ApiProperty()
-  email: string;
-
-  @IsString()
-  @IsNotEmpty()
-  @MinLength(6)
-  @ApiProperty()
-  password: string;
-
-  @IsNotEmpty()
-  @ApiProperty()
-  @IsEnum(UserGender)
+  @IsEnum(UserGender, {
+    message: 'gender must be a valid enum value (female, male)',
+  })
   gender: UserGender;
 
   @IsOptional()
@@ -43,4 +33,8 @@ export class UserBaseDto {
   @IsDate()
   @Type(() => Date) // Ensures proper transformation from string to Date
   dateOfBirth?: Date;
+
+  @IsOptional()
+  @ApiPropertyOptional()
+  contact?: ContactInfoDto;
 }
