+++ src/modules/user/dtos/user-base.dto.ts
@@ -1,16 +1,17 @@
 import { UserGender } from '@app-types/module.types';
-import { ApiProperty, ApiPropertyOptional } from '@nestjs/swagger';
+import { ApiProperty } from '@nestjs/swagger';
 import { Type } from 'class-transformer';
 import {
   IsString,
+  IsEmail,
   IsNotEmpty,
+  MinLength,
   IsOptional,
   IsDate,
   IsEnum,
 } from 'class-validator';
-import { ContactInfoDto } from './contact-info.dto';
 
-export class CreateUserDto {
+export class UserBaseDto {
   @IsString()
   @IsNotEmpty()
   @ApiProperty()
@@ -21,9 +22,22 @@ export class CreateUserDto {
   @ApiProperty()
   lastName: string;
 
+  @IsEmail()
   @IsNotEmpty()
   @ApiProperty()
-  @IsEnum(UserGender)
+  email: string;
+
+  @IsString()
+  @IsNotEmpty()
+  @MinLength(6)
+  @ApiProperty()
+  password: string;
+
+  @IsNotEmpty()
+  @ApiProperty()
+  @IsEnum(UserGender, {
+    message: 'gender must be a valid enum value (female, male)',
+  })
   gender: UserGender;
 
   @IsOptional()
@@ -31,8 +45,4 @@ export class CreateUserDto {
   @IsDate()
   @Type(() => Date) // Ensures proper transformation from string to Date
   dateOfBirth?: Date;
-
-  @IsOptional()
-  @ApiPropertyOptional()
-  contact?: ContactInfoDto;
 }
